backend/tmr-antigravity-v3/requirements/api.txt
# requirements/api.txt — FastAPI application + Celery workers

# ── Web framework ──────────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
uvloop>=0.19.0

# ── Database ───────────────────────────────────────────────────────────────────
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# ── Auth ───────────────────────────────────────────────────────────────────────
pyjwt>=2.8.0
bcrypt>=4.1.0
httpx>=0.27.0

# ── Security / Encryption ──────────────────────────────────────────────────────
cryptography>=43.0.0

# ── Task queue ─────────────────────────────────────────────────────────────────
celery[redis]>=5.4.0
redis>=5.0.0

# ── Monitoring ─────────────────────────────────────────────────────────────────
prometheus-fastapi-instrumentator>=7.0.0
sentry-sdk[fastapi,sqlalchemy]>=2.0.0
structlog>=24.0.0

# ── AI / LLM ───────────────────────────────────────────────────────────────────
# (HTTP calls only via httpx — no heavy LLM SDKs needed in API layer)

# ── Voice ──────────────────────────────────────────────────────────────────────
google-cloud-speech>=2.26.0
google-cloud-texttospeech>=2.17.0
pyaudio>=0.2.14
pyttsx3>=2.90           # offline TTS fallback
openai-whisper>=20231117  # offline STT fallback

# ── Cloud storage ──────────────────────────────────────────────────────────────
boto3>=1.35.0

# ── Content safety ─────────────────────────────────────────────────────────────
# (Perspective API called via httpx — no SDK needed)

# ── Utilities ──────────────────────────────────────────────────────────────────
python-multipart>=0.0.9    # file uploads
python-dotenv>=1.0.0
